poker-vision-detection-enricher/pyproject.toml
[project]
name = "poker-vision-detection-enricher"
version = "0.1.0"
description = "Detection enricher module for routing detections to classification, OCR, and spatial reasoning"
requires-python = ">=3.10"
dependencies = [
    "easyocr>=1.7.0",
    "fastapi>=0.111.0",
    "httpx>=0.27.0",
    "pillow>=10.0.0",
    "uvicorn[standard]>=0.29.0",
]

[tool.uv]
package = false
